add more documentation and refactoring
