+++ metrics/web/lib/common/presentation/model/filter.dart
@@ -1,0 +1,5 @@
+/// An abstract class that represents a filter.
+abstract class Filter<T> {
+  /// Applies the filter to the given list of data.
+  List<T> apply(List<T> list);
+}
