--- metrics/web/lib/dashboard/presentation/model/filter.dart
@@ -1,3 +1,0 @@
-abstract class Filter<T> {
-  List<T> apply(List<T> list);
-}
