+++ metrics/web/test/test_utils/project_metrics_notifier_stub.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/foundation.dart';
-import 'package:metrics/dashboard/presentation/model/filter.dart';
+import 'package:metrics/common/presentation/model/filter.dart';
 import 'package:metrics/dashboard/presentation/model/project_metrics_data.dart';
 import 'package:metrics/dashboard/presentation/state/project_metrics_notifier.dart';
 import 'package:metrics_core/metrics_core.dart';
